BAAI/pyproject.toml
[tool.poetry]
name = "dicta-embeddings-api"
version = "0.3.0"
description = "A FastAPI project for generating text embeddings and reranking using BAAI models."
authors = ["oznav"]
packages = [{ include = "*", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = ">=0.110.0"
uvicorn = ">=0.30.0"
python-dotenv = "^1.0.0"
pydantic = "^2.7.0"
numpy = "^1.26.0"
orjson = "^3.10.0"
llama-cpp-python = { version = "^0.3.4", source = "llama-cpp" }

[[tool.poetry.source]]
name = "llama-cpp"
url = "https://abetlen.github.io/llama-cpp-python/whl/cu124"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
bump_message = "chore: release $new_version"

[tool.flake8]
max-line-length = 120

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
files = ["src", "tests"]
strict = true
ignore_missing_imports = true

[tool.vulture]
paths = ["src", "tests"]
min_confidence = 70

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
branch = true
command_line = "-m pytest"
source = ["src"]
relative_files = true

[tool.codespell]
skip = 'src/assets/*,scripts/*'